creating new view with all projects and their enitity counts with json file while we wait for new service for entity count creation
