+++ src/app/components/navigation/navigation.html
@@ -19,6 +19,10 @@
     </md-list-item>
     <md-list-item ng-click="$ctrl.$state.go('dashboard')">
       <md-icon md-font-icon="fa fa-desktop fa-fw"></md-icon>
+      <p>All Projects</p>
+    </md-list-item>
+    <md-list-item ng-click="$ctrl.$state.go('dashboardExpedition')">
+      <md-icon md-font-icon="fa fa-list fa-fw"></md-icon>
       <p>Project Overview</p>
     </md-list-item>
   </md-list>
