+++ src/app/views/dashboard/dashboard.html
@@ -1,109 +1,72 @@
 <div layout-fill layout="column">
-  <h3 class="text-center page-header">{{$ctrl.currentProject.projectTitle}} Overview</h3>
-  <p class="text-center">{{ $ctrl.currentProject.description}}</p>
-  <p class="text-center">
-    <strong>Project owned by user: </strong> {{ $ctrl.currentProject.user.username}}
-    ({{$ctrl.currentProject.user.email}})
-  </p>
-
-  <div layout="row" class="loading-overlay" ng-show="$ctrl.loading" layout-align="center center">
+  <h2 class="text-center page-header">Project Dashboard</h2>
+  <div
+    layout="row"
+    layout-fill
+    layout-sm="column"
+    ng-show="$ctrl.loading"
+    layout-align="center center"
+  >
     <md-progress-circular md-mode="indeterminate"></md-progress-circular>
   </div>
-
-  <div layout="row" ng-show="!$ctrl.loading && $ctrl.totalItems" layout-align="center center">
-    <md-button
-      class="md-primary md-raised project-download-button"
-      ng-click="$ctrl.downloadCsv()"
-      layout="row"
-      layout-align="center center"
-    >
-      <!-- Download -->
-      <span>Project CSV Archive</span>
-      <md-icon ng-show="$ctrl.loadingExpedition !== 'project'" md-font-icon="fa fa-download fa-xs">
-      </md-icon>
-      <md-progress-circular
-        ng-show="$ctrl.loadingExpedition === 'project'"
-        md-mode="indeterminate"
-        md-diameter="18"
-        class="md-accent md-hue-3"
-      ></md-progress-circular>
-    </md-button>
-  </div>
-
-  <!-- <table class="table table-hover table-condensed text-center" ng-show="!$ctrl.loading"> -->
-  <table
-    class="table table-hover table-condensed text-center"
-    ng-show="$ctrl.displayResults.length > 0"
-  >
-    <thead>
-      <th class="text-center">Expedition Title</th>
-      <th class="text-center" ng-repeat="header in $ctrl.headers">
-        {{$ctrl.humanReadableHeader(header)}}
-      </th>
-      <th class="text-center">GUID</th>
-      <th></th>
-    </thead>
-
-    <tbody>
-      <tr
-        ng-repeat="expedition in $ctrl.displayResults track by expedition.expeditionCode"
-        ng-click="$ctrl.results && $ctrl.viewData(expedition.expeditionCode)"
+  <div class="text-center">
+    <p>
+      <md-button ui-sref="create-project" class="md-raised md-primary"
+        ><md-icon md-font-icon="fa fa-plus fa-fw"></md-icon><span>New Project</span></md-button
       >
-        <td>{{ expedition.expeditionTitle}}</td>
-        <td ng-repeat="header in $ctrl.headers">{{ expedition[header] }}</td>
-        <td ng-show="expedition.expeditionIdentifier">
-          <!--<a href="https://n2t.net/{{expedition.expeditionIdentifier}}">https://n2t.net/{{ expedition.expeditionIdentifier}} </a>-->
-          https://n2t.net/{{ expedition.expeditionIdentifier}}
-        </td>
-        <td ng-show="!expedition.expeditionIdentifier"></td>
-        <td ng-show="$ctrl.menuOptions(expedition).length == 0"></td>
-        <td ng-show="$ctrl.menuOptions(expedition).length > 0" class="md-dense action-buttons">
-          <!--<md-button aria-label="View expedition data" ng-click="$ctrl.viewData(expedition.expeditionCode)"
-            class="md-primary md-raised query-button">
-            <md-icon md-font-icon="fa fa-map-marker fa-sm"></md-icon>
-          </md-button>-->
-          <md-menu ng-show="$ctrl.results" class="md-dense" md-position-mode="target-right target">
-            <md-button
-              aria-label="Open expedition download menu"
-              ng-click="$mdMenu.open($event)"
-              class="md-primary download-button"
-            >
-              <!-- Download -->
-              <md-icon
-                ng-show="$ctrl.loadingExpedition !== expedition.expeditionCode"
-                md-font-icon="fa fa-download fa-xs"
-              ></md-icon>
-              <md-progress-circular
-                ng-show="$ctrl.loadingExpedition === expedition.expeditionCode"
-                md-mode="indeterminate"
-                md-diameter="18"
-                class="md-accent md-hue-3"
-              >
-              </md-progress-circular>
-            </md-button>
-            <md-menu-content width="2" class="md-dense download-menu-content">
-              <md-menu-item
-                ng-repeat="opt in $ctrl.menuOptions(expedition) track by expedition.expeditionCode + opt.name"
-              >
-                <md-button ng-click="opt.fn(expedition.expeditionCode)">{{opt.name}}</md-button>
-              </md-menu-item>
-            </md-menu-content>
-          </md-menu>
-        </td>
-      </tr>
-    </tbody>
-  </table>
-
-  <div class="text-center" ng-show="$ctrl.totalItems > 0">
-    <ul
-      uib-pagination
-      total-items="$ctrl.totalItems"
-      ng-model="$ctrl.currentPage"
-      ng-change="$ctrl.pageChanged()"
-      max-size="10"
-      force-ellipses="true"
-      boundary-link-numbers="true"
-      items-per-page="$ctrl.itemsPerPage"
-    ></ul>
+    </p>
+    <p>
+      Click on project title to view its expeditions and entity counts
+    </p>
   </div>
+  <div ng-if="$ctrl.currentUser">
+    <h3>Private Projects for {{ $ctrl.currentUser.username }}</h3>
+    <table class="table table-hover table-condensed">
+      <thead>
+        <th class="text-center">Project Title</th>
+        <th class="text-center" ng-repeat="entity in $ctrl.entities">
+          {{ entity }}
+        </th>
+      </thead>
+
+      <tr ng-repeat="data in $ctrl.projectsAndTheirEntityCounts">
+        <!-- 
+        ng-show="data.user.username === $ctrl.currentUser.username" 
+        the service call for projects entity counts will need to include the user, or we redisign the table 
+        -->
+        <td>
+          <a href="#" ng-click="$ctrl.showExpeditionsDetail(project)">{{ data.project_title }}</a>
+        </td>
+        <td class="text-center">{{ data.samples }}</td>
+        <td class="text-center">{{ data.events }}</td>
+        <td class="text-center">{{ data.tissues }}</td>
+        <td class="text-center">{{ data.event_photos }}</td>
+        <td class="text-center">{{ data.fasta }}</td>
+        <td class="text-center">{{ data.fastq }}</td>
+        <td></td>
+      </tr>
+    </table>
+  </div>
+
+  <h3>Public Projects</h3>
+  <table class="table table-hover table-condensed">
+    <thead>
+      <th>Project Title</th>
+      <th ng-repeat="entity in $ctrl.entities">
+        {{ entity }}
+      </th>
+    </thead>
+    <tr ng-repeat="data in $ctrl.projectsAndTheirEntityCounts">
+      <!--  ng-if="$ctrl.currentUser.username !== data.user.username" -->
+      <td>
+        <a href="#" ng-click="$ctrl.showExpeditionsDetail(project)">{{ data.project_title }}</a>
+      </td>
+      <td class="text-center">{{ data.samples }}</td>
+      <td class="text-center">{{ data.events }}</td>
+      <td class="text-center">{{ data.tissues }}</td>
+      <td class="text-center">{{ data.event_photos }}</td>
+      <td class="text-center">{{ data.fasta }}</td>
+      <td class="text-center">{{ data.fastq }}</td>
+    </tr>
+  </table>
 </div>
