+++ src/app/views/dashboard/index.js
@@ -2,8 +2,11 @@ import angular from 'angular';
 
 import routing from './dashboard.routes';
 import dashboard from './dashboard.component';
+import dashboardExpedition from './dashboard-expedition.component';
+import projectService from '../../services/project.service';
 
 export default angular
-  .module('fims.dashboard', [])
+  .module('fims.dashboard', [projectService])
   .run(routing)
-  .component('fimsDashboard', dashboard).name;
+  .component('fimsDashboard', dashboard)
+  .component('fimsDashboardExpedition', dashboardExpedition).name;
