+++ src/app/views/expeditions/expeditions.html
@@ -9,7 +9,7 @@
       and may not represent a comprehensive list of expeditions for this project, which can be found
       under
       <br />
-      <md-button href ui-sref="dashboard">
+      <md-button href ui-sref="dashboardExpedition">
         <md-icon md-font-icon="fa fa-desktop fa-fw"></md-icon>
         Project Overview
       </md-button>
